requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-huggingface==0.0.6
pypdf==3.17.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
streamlit==1.29.0
openai>=1.6.1,<2.0.0 